Re-org base api client/models/resources into modules
